tonights  fix

+++ DLL/LouDLLs/Codecs/Bmp/BmpClips.c
@@ -5,7 +5,7 @@ void
 AnnyaPaintClipWithBmp(
     HANDLE  Bitmap,
     HANDLE  Cliph,
-    INT64  X, INT64  Y,
+    size_t  X, size_t  Y,
     size_t  ScalingX,
     size_t  ScalingY
 ){
