+++ include/kernel/DRSD.h
@@ -416,10 +416,10 @@ typedef struct _DRSD_PLANE{
     size_t                          BlueShift;
     size_t                          GreenShift;
     bool                            IsDirty;
-    size_t                          DirtyX;
-    size_t                          DirtyY;
-    size_t                          DirtyWidth;
-    size_t                          DirtyHeight;
+    INT64                           DirtyX;
+    INT64                           DirtyY;
+    INT64                           DirtyWidth;
+    INT64                           DirtyHeight;
 }DRSD_PLANE, * PDRSD_PLANE;
 
 typedef enum{
